+++ Graphics/Slicer/Math/Skeleton/Concave.hs
@@ -28,7 +28,7 @@
 
 module Graphics.Slicer.Math.Skeleton.Concave (averageNodes, skeletonOfConcaveRegion, findINodes, makeENode, makeENodes, eNodesOfOutsideContour) where
 
-import Prelude (Eq, Show, Bool(True, False), Either(Left, Right), String, Ord, Ordering(GT,LT), all, concatMap, notElem, otherwise, ($), (>), (<=), (<$>), (==), (/=), error, (&&), fst, (<>), show, not, max, compare, uncurry, null, (||), min, snd, filter, zip, any, (*), (+), Int, (.), (-), mempty, realToFrac)
+import Prelude (Eq, Show, Bool(True, False), Either(Left, Right), String, Ord, Ordering(GT,LT), all, concatMap, notElem, otherwise, ($), (>), (<=), (<$>), (==), (/=), error, (&&), fst, (<>), show, not, max, compare, uncurry, null, (||), min, snd, filter, zip, any, (*), (+), Int, (.), (-), mempty, )
 
 import Prelude as PL (head, last, tail, init)
 
@@ -688,10 +688,10 @@ skeletonOfNodes connectedLoop origSegSets inSegSets iNodes =
       then removeFoundINodesOf shortestMixedPairs
       else []
       where
+        -- | Filter out any node pairs in our list that contain an INode from the iPairsFound list.
         removeFoundINodesOf :: [(ENode, INode)] -> [(ENode, INode)]
         removeFoundINodesOf = filter withoutINodes
           where
-            -- filter out any nodes in our list that are in the iPairsFound result.
             withoutINodes (_,myINode) = myINode `notElem` ((fst <$> iPairsFound) <> (snd <$> iPairsFound))
 
     ePairsFound =
@@ -699,7 +699,7 @@ skeletonOfNodes connectedLoop origSegSets inSegSets iNodes =
       then removeFoundENodesOf $ shortestNeighboringPairs $ mapWithFollower (,) eNodes
       else []
       where
-        -- filter out any nodes in our list, that are in the mixedPairsFound result.
+        -- | Filter out any node pairs in our list that contain an ENode from the mixedPairsFound list.
         removeFoundENodesOf :: [(ENode, ENode)] -> [(ENode, ENode)]
         removeFoundENodesOf = filter withoutENodes
           where
@@ -741,9 +741,9 @@ skeletonOfNodes connectedLoop origSegSets inSegSets iNodes =
       | isAntiParallel (outAndErrOf n1) (outAndErrOf n2) = error $ "Cannot get the average of nodes if their outputs never intersect!\n" <> errorLen3
       | isCollinear (outAndErrOf n1) (outAndErrOf n2) = error $ "Cannot (yet) handle two input plines that are collinear.\n" <> errorLen3
       | nodesAreAntiCollinear n1 n2 = error $ "Cannot (yet) handle two input plines that are collinear.\n" <> errorLen3
-      | n1Distance <= realToFrac (ulpVal n1Err) = error $ "intersection is AT the point of n1!\n" <> show n1Distance <> "\n" <> show n2Distance <> "\n" <> show intersectionPoint <> "\n" <> show n1 <> "\n" <> show n2 <> "\n" <> errorLen3
-      | n2Distance <= realToFrac (ulpVal n2Err) = error $ "intersection is AT the point of n2!\n" <> show n1Distance <> "\n" <> show n2Distance <> "\n" <> show intersectionPoint <> "\n" <> show n1 <> "\n" <> show n2 <> "\n" <> errorLen3
-      | n1Distance > realToFrac (ulpVal n1Err) && n2Distance > realToFrac (ulpVal n2Err) = averageNodes n1 n2
+      | n1Distance <= ulpVal n1Err = error $ "intersection is AT the point of n1!\n" <> show n1Distance <> "\n" <> show n2Distance <> "\n" <> show intersectionPoint <> "\n" <> show n1 <> "\n" <> show n2 <> "\n" <> errorLen3
+      | n2Distance <= ulpVal n2Err = error $ "intersection is AT the point of n2!\n" <> show n1Distance <> "\n" <> show n2Distance <> "\n" <> show intersectionPoint <> "\n" <> show n1 <> "\n" <> show n2 <> "\n" <> errorLen3
+      | n1Distance > ulpVal n1Err && n2Distance > ulpVal n2Err = averageNodes n1 n2
       | otherwise = error $ "found node too close:\n"
                     <> show n1 <> "\n"
                     <> show n2 <> "\n"
@@ -816,8 +816,8 @@ skeletonOfNodes connectedLoop origSegSets inSegSets iNodes =
     intersectsInPoint :: (Arcable a, Pointable a, Arcable b, Pointable b) => a -> b -> Bool
     intersectsInPoint node1 node2
       | hasArc node1 && hasArc node2 = not (noIntersection (outAndErrOf node1) (outAndErrOf node2))
-                                       && not (dist1 <= realToFrac (ulpVal dist1Err))
-                                       && not (dist2 <= realToFrac (ulpVal dist2Err))
+                                       && not (dist1 <= ulpVal dist1Err)
+                                       && not (dist2 <= ulpVal dist2Err)
       | otherwise                    = error $ "cannot intersect a node with no output:\nNode1: " <> show node1 <> "\nNode2: " <> show node2 <> "\nnodes: " <> show iNodes <> "\n"
       where
         (dist1, (_,_, dist1Err)) = distance2PP (intersectionOf (outAndErrOf node1) (outAndErrOf node2)) (pPointAndErrOf node1)
